sleep: suppress static inline in sleep.h
move inline function in sleep.c to suppress inline
declaration in irq.h
Allow the use of libmetal in some environments such as
FFI (linking from rust). Functions defined as #defines,
or as static inlines in headers are useless in this context.
The compiled library simply does not contain these symbols.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>